settings.py moved into package
